WBT: Move more code from BuildTestBase to the respective provider, and test classes
