+++ src/mono/wasm/Wasm.Build.Tests/BuildTestBase.cs
@@ -48,9 +48,7 @@ namespace Wasm.Build.Tests
         // FIXME: use an envvar to override this
         protected static int s_defaultPerTestTimeoutMs = s_isWindows ? 30 * 60 * 1000 : 15 * 60 * 1000;
         public static BuildEnvironment s_buildEnv;
-        private const string s_runtimePackPathPattern = "\\*\\* MicrosoftNetCoreAppRuntimePackDir : '([^ ']*)'";
         private const string s_nugetInsertionTag = "<!-- TEST_RESTORE_SOURCES_INSERTION_LINE -->";
-        private static Regex s_runtimePackPathRegex;
 
         public static bool IsUsingWorkloads => s_buildEnv.IsWorkload;
         public static bool IsNotUsingWorkloads => !s_buildEnv.IsWorkload;
@@ -58,8 +56,6 @@ namespace Wasm.Build.Tests
         public static string GetNuGetConfigPathFor(string targetFramework) =>
             Path.Combine(BuildEnvironment.TestDataPath, "nuget8.config"); // for now - we are still using net7, but with
                                                                           // targetFramework == "net7.0" ? "nuget7.config" : "nuget8.config");
-
-        public const string WebcilInWasmExtension = ".wasm";
 
         public TProvider GetProvider<TProvider>() where TProvider : ProjectProviderBase
             => (TProvider)_providerOfBaseType;
@@ -75,7 +71,6 @@ namespace Wasm.Build.Tests
             try
             {
                 s_buildEnv = new BuildEnvironment();
-                s_runtimePackPathRegex = new Regex(s_runtimePackPathPattern);
 
                 s_skipProjectCleanup = !string.IsNullOrEmpty(EnvironmentVariables.SkipProjectCleanup) && EnvironmentVariables.SkipProjectCleanup == "1";
 
@@ -357,160 +352,6 @@ namespace Wasm.Build.Tests
             return buildArgs with { ProjectFileContents = projectContents };
         }
 
-        public (string projectDir, string buildOutput) BuildTemplateProject(BuildArgs buildArgs,
-                                  string id,
-                                  BuildProjectOptions buildProjectOptions,
-                                  AssertTestMainJsAppBundleOptions? assertAppBundleOptions = null)
-        {
-            StringBuilder buildCmdLine = new();
-            buildCmdLine.Append(buildProjectOptions.Publish ? "publish" : "build");
-
-            string logFilePath = Path.Combine(s_buildEnv.LogRootPath, $"{id}.binlog");
-            _testOutput.WriteLine($"-------- Building ---------");
-            _testOutput.WriteLine($"Binlog path: {logFilePath}");
-            buildCmdLine.Append($" -c {buildArgs.Config} -bl:{logFilePath} {buildArgs.ExtraBuildArgs}");
-
-            if (buildProjectOptions.Publish && buildProjectOptions.BuildOnlyAfterPublish)
-                buildCmdLine.Append(" -p:WasmBuildOnlyAfterPublish=true");
-
-            CommandResult res = new DotNetCommand(s_buildEnv, _testOutput)
-                                    .WithWorkingDirectory(_projectDir!)
-                                    .WithEnvironmentVariables(buildProjectOptions.ExtraBuildEnvironmentVariables)
-                                    .ExecuteWithCapturedOutput(buildCmdLine.ToString());
-            if (buildProjectOptions.ExpectSuccess)
-                res.EnsureSuccessful();
-            else
-                Assert.NotEqual(0, res.ExitCode);
-
-            if (buildProjectOptions.UseCache)
-                _buildContext.CacheBuild(buildArgs, new BuildProduct(_projectDir!, logFilePath, true, res.Output));
-
-            AssertRuntimePackPath(res.Output, buildProjectOptions.TargetFramework ?? DefaultTargetFramework);
-            string bundleDir = Path.Combine(GetBinDir(config: buildArgs.Config, targetFramework: buildProjectOptions.TargetFramework ?? DefaultTargetFramework), "AppBundle");
-
-            assertAppBundleOptions ??= new AssertTestMainJsAppBundleOptions(
-                                            BundleDir: bundleDir,
-                                            ProjectName: buildArgs.ProjectName,
-                                            Config: buildArgs.Config,
-                                            MainJS: buildProjectOptions.MainJS ?? "test-main.js",
-                                            HasV8Script: buildProjectOptions.HasV8Script,
-                                            GlobalizationMode: buildProjectOptions.GlobalizationMode,
-                                            PredefinedIcudt: buildProjectOptions.PredefinedIcudt ?? "",
-                                            UseWebcil: UseWebcil,
-                                            IsBrowserProject: buildProjectOptions.IsBrowserProject,
-                                            IsPublish: buildProjectOptions.Publish);
-            AssertBasicAppBundle(assertAppBundleOptions);
-
-            return (_projectDir!, res.Output);
-        }
-
-        public (string projectDir, string buildOutput) BuildProject(BuildArgs buildArgs,
-                                  string id,
-                                  BuildProjectOptions options)
-        {
-            string msgPrefix = options.Label != null ? $"[{options.Label}] " : string.Empty;
-            if (options.UseCache && _buildContext.TryGetBuildFor(buildArgs, out BuildProduct? product))
-            {
-                _testOutput.WriteLine($"Using existing build found at {product.ProjectDir}, with build log at {product.LogFile}");
-
-                if (!product.Result)
-                    throw new XunitException($"Found existing build at {product.ProjectDir}, but it had failed. Check build log at {product.LogFile}");
-                _projectDir = product.ProjectDir;
-
-                // use this test's id for the run logs
-                _logPath = Path.Combine(s_buildEnv.LogRootPath, id);
-                return (_projectDir, product.BuildOutput);
-            }
-
-            if (options.CreateProject)
-            {
-                InitPaths(id);
-                InitProjectDir(_projectDir);
-                options.InitProject?.Invoke();
-
-                File.WriteAllText(Path.Combine(_projectDir, $"{buildArgs.ProjectName}.csproj"), buildArgs.ProjectFileContents);
-                File.Copy(
-                    Path.Combine(
-                        AppContext.BaseDirectory,
-                        string.IsNullOrEmpty(options.TargetFramework) || options.TargetFramework == "net8.0"
-                            ? "test-main.js"
-                            : "data/test-main-7.0.js"
-                    ),
-                    Path.Combine(_projectDir, "test-main.js")
-                );
-
-                File.WriteAllText(Path.Combine(_projectDir!, "index.html"), @"<html><body><script type=""module"" src=""test-main.js""></script></body></html>");
-            }
-            else if (_projectDir is null)
-            {
-                throw new Exception("_projectDir should be set, to use options.createProject=false");
-            }
-
-            StringBuilder sb = new();
-            sb.Append(options.Publish ? "publish" : "build");
-            if (options.Publish && options.BuildOnlyAfterPublish)
-                sb.Append(" -p:WasmBuildOnlyAfterPublish=true");
-            sb.Append($" {s_buildEnv.DefaultBuildArgs}");
-
-            sb.Append($" /p:Configuration={buildArgs.Config}");
-
-            string logFileSuffix = options.Label == null ? string.Empty : options.Label.Replace(' ', '_');
-            string logFilePath = Path.Combine(_logPath, $"{buildArgs.ProjectName}{logFileSuffix}.binlog");
-            _testOutput.WriteLine($"-------- Building ---------");
-            _testOutput.WriteLine($"Binlog path: {logFilePath}");
-            sb.Append($" /bl:\"{logFilePath}\" /nologo");
-            sb.Append($" /v:{options.Verbosity ?? "minimal"}");
-            if (buildArgs.ExtraBuildArgs != null)
-                sb.Append($" {buildArgs.ExtraBuildArgs} ");
-
-            _testOutput.WriteLine($"Building {buildArgs.ProjectName} in {_projectDir}");
-
-            (int exitCode, string buildOutput) result;
-            try
-            {
-                var envVars = s_buildEnv.EnvVars;
-                if (options.ExtraBuildEnvironmentVariables is not null)
-                {
-                    envVars = new Dictionary<string, string>(s_buildEnv.EnvVars);
-                    foreach (var kvp in options.ExtraBuildEnvironmentVariables!)
-                        envVars[kvp.Key] = kvp.Value;
-                }
-                envVars["NUGET_PACKAGES"] = _nugetPackagesDir;
-                result = AssertBuild(sb.ToString(), id, expectSuccess: options.ExpectSuccess, envVars: envVars);
-
-                // check that we are using the correct runtime pack!
-
-                if (options.ExpectSuccess && options.AssertAppBundle)
-                {
-                    AssertRuntimePackPath(result.buildOutput, options.TargetFramework ?? DefaultTargetFramework);
-
-                    string bundleDir = Path.Combine(GetBinDir(config: buildArgs.Config, targetFramework: options.TargetFramework ?? DefaultTargetFramework), "AppBundle");
-                    AssertBasicAppBundle(new AssertTestMainJsAppBundleOptions(
-                                            BundleDir: bundleDir,
-                                            ProjectName: buildArgs.ProjectName,
-                                            Config: buildArgs.Config,
-                                            MainJS: options.MainJS ?? "test-main.js",
-                                            HasV8Script: options.HasV8Script,
-                                            GlobalizationMode: options.GlobalizationMode,
-                                            PredefinedIcudt: options.PredefinedIcudt ?? "",
-                                            UseWebcil: UseWebcil,
-                                            IsBrowserProject: options.IsBrowserProject,
-                                            IsPublish: options.Publish));
-                }
-
-                if (options.UseCache)
-                    _buildContext.CacheBuild(buildArgs, new BuildProduct(_projectDir, logFilePath, true, result.buildOutput));
-
-                return (_projectDir, result.buildOutput);
-            }
-            catch (Exception ex)
-            {
-                if (options.UseCache)
-                    _buildContext.CacheBuild(buildArgs, new BuildProduct(_projectDir, logFilePath, false, $"The build attempt resulted in exception: {ex}."));
-                throw;
-            }
-        }
-
         protected static string GetNuGetConfigWithLocalPackagesPath(string templatePath, string localNuGetsPath)
         {
             string contents = File.ReadAllText(templatePath);
@@ -518,173 +359,6 @@ namespace Wasm.Build.Tests
                 throw new Exception($"Could not find {s_nugetInsertionTag} in {templatePath}");
 
             return contents.Replace(s_nugetInsertionTag, $@"<add key=""nuget-local"" value=""{localNuGetsPath}"" />");
-        }
-
-        public string CreateWasmTemplateProject(string id, string template = "wasmbrowser", string extraArgs = "", bool runAnalyzers = true)
-        {
-            InitPaths(id);
-            InitProjectDir(_projectDir, addNuGetSourceForLocalPackages: true);
-
-            File.WriteAllText(Path.Combine(_projectDir, "Directory.Build.props"), "<Project />");
-            File.WriteAllText(Path.Combine(_projectDir, "Directory.Build.targets"),
-                """
-                <Project>
-                  <Target Name="PrintRuntimePackPath" BeforeTargets="Build">
-                      <Message Text="** MicrosoftNetCoreAppRuntimePackDir : '@(ResolvedRuntimePack -> '%(PackageDirectory)')'" Importance="High" Condition="@(ResolvedRuntimePack->Count()) > 0" />
-                  </Target>
-                </Project>
-                """);
-
-            new DotNetCommand(s_buildEnv, _testOutput, useDefaultArgs: false)
-                    .WithWorkingDirectory(_projectDir!)
-                    .ExecuteWithCapturedOutput($"new {template} {extraArgs}")
-                    .EnsureSuccessful();
-
-            string projectfile = Path.Combine(_projectDir!, $"{id}.csproj");
-            string extraProperties = string.Empty;
-            extraProperties += "<TreatWarningsAsErrors>true</TreatWarningsAsErrors>";
-            if (runAnalyzers)
-                extraProperties += "<RunAnalyzers>true</RunAnalyzers>";
-            if (!UseWebcil)
-                extraProperties += "<WasmEnableWebcil>false</WasmEnableWebcil>";
-
-            // TODO: Can be removed after updated templates propagate in.
-            string extraItems = string.Empty;
-            if (template == "wasmbrowser")
-                extraItems += "<WasmExtraFilesToDeploy Include=\"main.js\" />";
-            else
-                extraItems += "<WasmExtraFilesToDeploy Include=\"main.mjs\" />";
-
-            AddItemsPropertiesToProject(projectfile, extraProperties, extraItems);
-
-            return projectfile;
-        }
-
-        static void AssertRuntimePackPath(string buildOutput, string targetFramework)
-        {
-            var match = s_runtimePackPathRegex.Match(buildOutput);
-            if (!match.Success || match.Groups.Count != 2)
-                throw new XunitException($"Could not find the pattern in the build output: '{s_runtimePackPathPattern}'.{Environment.NewLine}Build output: {buildOutput}");
-
-            string expectedRuntimePackDir = s_buildEnv.GetRuntimePackDir(targetFramework);
-            string actualPath = match.Groups[1].Value;
-            if (string.Compare(actualPath, expectedRuntimePackDir) != 0)
-                throw new XunitException($"Runtime pack path doesn't match.{Environment.NewLine}Expected: '{expectedRuntimePackDir}'{Environment.NewLine}Actual:   '{actualPath}'");
-        }
-
-        private void AssertBasicAppBundle(AssertTestMainJsAppBundleOptions options)
-        {
-            new TestMainJsProjectProvider(_testOutput, _projectDir)
-                    .FindAndAssertDotnetFiles(
-                        Path.Combine(options.BundleDir, "_framework"),
-                        isPublish: options.IsPublish,
-                        expectFingerprintOnDotnetJs: false,
-                        runtimeType: RuntimeVariant.SingleThreaded);
-
-            var filesToExist = new List<string>()
-            {
-                options.MainJS,
-                "_framework/blazor.boot.json",
-                "_framework/dotnet.js.map",
-                "_framework/dotnet.runtime.js.map",
-            };
-
-            if (options.IsBrowserProject)
-                filesToExist.Add("index.html");
-
-            TestUtils.AssertFilesExist(options.BundleDir, filesToExist);
-
-            TestUtils.AssertFilesExist(options.BundleDir, new[] { "run-v8.sh" }, expectToExist: options.HasV8Script);
-            AssertIcuAssets();
-
-            string managedDir = Path.Combine(options.BundleDir, "_framework");
-            string bundledMainAppAssembly =
-                options.UseWebcil ? $"{options.ProjectName}{WebcilInWasmExtension}" : $"{options.ProjectName}.dll";
-            TestUtils.AssertFilesExist(managedDir, new[] { bundledMainAppAssembly });
-
-            bool is_debug = options.Config == "Debug";
-            if (is_debug)
-            {
-                // Use cecil to check embedded pdb?
-                // AssertFilesExist(managedDir, new[] { $"{projectName}.pdb" });
-
-                //FIXME: um.. what about these? embedded? why is linker omitting them?
-                //foreach (string file in Directory.EnumerateFiles(managedDir, "*.dll"))
-                //{
-                //string pdb = Path.ChangeExtension(file, ".pdb");
-                //Assert.True(File.Exists(pdb), $"Could not find {pdb} for {file}");
-                //}
-            }
-
-            void AssertIcuAssets()
-            {
-                bool expectEFIGS = false;
-                bool expectCJK = false;
-                bool expectNOCJK = false;
-                bool expectFULL = false;
-                bool expectHYBRID = false;
-                switch (options.GlobalizationMode)
-                {
-                    case GlobalizationMode.Invariant:
-                        break;
-                    case GlobalizationMode.FullIcu:
-                        expectFULL = true;
-                        break;
-                    case GlobalizationMode.Hybrid:
-                        expectHYBRID = true;
-                        break;
-                    case GlobalizationMode.PredefinedIcu:
-                        if (string.IsNullOrEmpty(options.PredefinedIcudt))
-                            throw new ArgumentException("WasmBuildTest is invalid, value for predefinedIcudt is required when GlobalizationMode=PredefinedIcu.");
-                        TestUtils.AssertFilesExist(options.BundleDir, new[] { Path.Combine("_framework", options.PredefinedIcudt) }, expectToExist: true);
-                        // predefined ICU name can be identical with the icu files from runtime pack
-                        switch (options.PredefinedIcudt)
-                        {
-                            case "icudt.dat":
-                                expectFULL = true;
-                                break;
-                            case "icudt_EFIGS.dat":
-                                expectEFIGS = true;
-                                break;
-                            case "icudt_CJK.dat":
-                                expectCJK = true;
-                                break;
-                            case "icudt_no_CJK.dat":
-                                expectNOCJK = true;
-                                break;
-                        }
-                        break;
-                    default:
-                        // icu shard chosen based on the locale
-                        expectCJK = true;
-                        expectEFIGS = true;
-                        expectNOCJK = true;
-                        break;
-                }
-
-                var frameworkDir = Path.Combine(options.BundleDir, "_framework");
-                TestUtils.AssertFilesExist(frameworkDir, new[] { "icudt.dat" }, expectToExist: expectFULL);
-                TestUtils.AssertFilesExist(frameworkDir, new[] { "icudt_EFIGS.dat" }, expectToExist: expectEFIGS);
-                TestUtils.AssertFilesExist(frameworkDir, new[] { "icudt_CJK.dat" }, expectToExist: expectCJK);
-                TestUtils.AssertFilesExist(frameworkDir, new[] { "icudt_no_CJK.dat" }, expectToExist: expectNOCJK);
-                TestUtils.AssertFilesExist(frameworkDir, new[] { "icudt_hybrid.dat" }, expectToExist: expectHYBRID);
-            }
-        }
-
-        protected static void AssertDotNetJsSymbols(string bundleDir, bool fromRuntimePack, string targetFramework)
-            => TestUtils.AssertFile(Path.Combine(s_buildEnv.GetRuntimeNativeDir(targetFramework), "dotnet.native.js.symbols"),
-                            Path.Combine(bundleDir, "_framework/dotnet.native.js.symbols"),
-                            same: fromRuntimePack);
-
-        protected (int exitCode, string buildOutput) AssertBuild(string args, string label = "build", bool expectSuccess = true, IDictionary<string, string>? envVars = null, int? timeoutMs = null)
-        {
-            var result = RunProcess(s_buildEnv.DotNet, _testOutput, args, workingDir: _projectDir, label: label, envVars: envVars, timeoutMs: timeoutMs ?? s_defaultPerTestTimeoutMs);
-            if (expectSuccess && result.exitCode != 0)
-                throw new XunitException($"Build process exited with non-zero exit code: {result.exitCode}");
-            if (!expectSuccess && result.exitCode == 0)
-                throw new XunitException($"Build should have failed, but it didn't. Process exited with exitCode : {result.exitCode}");
-
-            return result;
         }
 
         public string FindBlazorBinFrameworkDir(string config, bool forPublish, string framework = DefaultTargetFrameworkForBlazor)
